+++ test/spec/collection/liquibase.rb
@@ -1,0 +1,4 @@
+shared_examples 'collection::liquibase' do
+    include_examples 'liquibase::liquibase'
+end
+