+++ test/spec/shared/liquibase/liquibase.rb
@@ -1,0 +1,5 @@
+shared_examples 'liquibase::liquibase' do
+    describe command('liquibase --version') do
+        its(:exit_status) { should eq 0 }
+    end
+end